requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cli==0.0.6
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
psycopg2==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
python-dotenv==1.0.1
python-environ==0.4.54
python-multipart==0.0.19
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.12.0
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.2
starlette==0.41.3
tabulate==0.9.0
typer==0.15.1
typing_extensions==4.12.2
uvicorn==0.32.1
watchfiles==1.0.0
websockets==14.1
yoyo-migrations==9.0.0
zipp==3.21.0
